+++ pkg/streamingpromql/optimize/plan/commonsubexpressionelimination/optimization_pass.go
@@ -136,7 +136,8 @@ func (e *OptimizationPass) accumulatePath(soFar path) []path {
 }
 
 // ShouldSkipChild determines if a child node should be skipped during common subexpression elimination.
-// Currently this is only used to skip the 2nd argument to info function calls, as that is some parameters specified by the user and not a direct query.
+// Currently this is only used to skip the 2nd argument to info function calls, as we don't want to
+// deduplicate these since those are parameters specified by the user and not a direct query of the info metric.
 func (e *OptimizationPass) ShouldSkipChild(node planning.Node, childIdx int) bool {
 	functionCall, ok := node.(*core.FunctionCall)
 	if !ok {